media - migration re run option

+++ src/bp-media/bp-media-functions.php
@@ -1010,16 +1010,6 @@ function bp_media_handle_sideload( $file_array, $post_data = array() ) {
 }
 
 /**
- * Update import status for media plugin import
- *
- * @since BuddyBoss 1.0.0
- */
-function bp_media_update_import_status() {
-	update_option( 'bp_media_import_status', 'done' );
-	update_option( 'bp_media_import_status_timestamp', bp_core_current_time() );
-}
-
-/**
  * Import BuddyBoss Media plugin db tables into Media Component
  *
  * @since BuddyBoss 1.0.0
@@ -1036,7 +1026,7 @@ function bp_media_import_buddyboss_media_tables() {
 	update_option( 'bp_media_import_total_media', $total_media );
 	update_option( 'bp_media_import_total_albums', $total_albums );
 
-	$albums_done = get_option( 'bp_media_import_albums_done', 0 );
+	$albums_done      = get_option( 'bp_media_import_albums_done', 0 );
 	$run_albums_query = $albums_done != $total_albums;
 
 	if ( $run_albums_query ) {
@@ -1092,8 +1082,8 @@ function bp_media_import_buddyboss_media_tables() {
 
 	if ( ! $run_albums_query ) {
 
-		$media_done = get_option( 'bp_media_import_media_done', 0 );
-		$album_ids = get_option( 'bp_media_import_albums_ids', array() );
+		$media_done         = get_option( 'bp_media_import_media_done', 0 );
+		$album_ids          = get_option( 'bp_media_import_albums_ids', array() );
 		$imported_media_ids = get_option( 'bp_media_import_media_ids', array() );
 
 		$medias = $wpdb->get_results( "SELECT * FROM {$buddyboss_media_table} LIMIT 100 OFFSET {$media_done}" );
@@ -1130,7 +1120,7 @@ function bp_media_import_buddyboss_media_tables() {
 					}
 				}
 
-				if ( ! empty( $media->upload_date ) ) {
+				if ( ! empty( $media->upload_date ) && '0000-00-00 00:00:00' != $media->upload_date ) {
 					$date_created = $media->upload_date;
 				} else if ( ! empty( $media->upload_date ) && '0000-00-00 00:00:00' == $media->upload_date && ! empty( $attachment_id ) ) {
 					$date_created = get_the_date( $attachment_id );
@@ -1154,37 +1144,42 @@ function bp_media_import_buddyboss_media_tables() {
 
 				$media_args['privacy'] = $privacy;
 
-				if ( ! empty( $activity_id ) && bp_is_active( 'activity' ) ) {
+				if ( bp_is_active( 'activity' ) ) {
 
-					$activity = new BP_Activity_Activity( $activity_id );
+					$activity_args = array(
+						'user_id'       => $user_id,
+						'recorded_time' => $date_created,
+						'hide_sitewide' => true,
+						'privacy'       => 'media',
+						'type'          => 'activity_update',
+						'component'     => buddypress()->activity->id,
+					);
 
-					if ( ! empty( $activity->id ) ) {
+					if ( ! empty( $activity_id ) ) {
 
-						$activity_args = array(
-							'user_id'       => $user_id,
-							'recorded_time' => $activity->date_recorded,
-							'hide_sitewide' => true,
-							'privacy'       => 'media',
-							'type'          => 'activity_update',
-							'component'     => buddypress()->activity->id,
-						);
+						$activity = new BP_Activity_Activity( $activity_id );
 
-						if ( 'groups' == $activity->component ) {
-							$media_args['group_id'] = $activity->item_id;
+						if ( ! empty( $activity->id ) ) {
 
-							$activity_args['component'] = buddypress()->groups->id;
-							$activity_args['item_id']   = $activity->item_id;
+							$activity_args['recorded_time'] = $activity->date_recorded;
+
+							if ( 'groups' == $activity->component ) {
+								$media_args['group_id'] = $activity->item_id;
+
+								$activity_args['component'] = buddypress()->groups->id;
+								$activity_args['item_id']   = $activity->item_id;
+							}
 						}
+					}
 
-						// make an activity for the media
-						$sub_activity_id = bp_activity_add( $activity_args );
+					// make an activity for the media
+					$sub_activity_id = bp_activity_add( $activity_args );
 
-						if ( $sub_activity_id ) {
-							// update activity meta
-							bp_activity_update_meta( $sub_activity_id, 'bp_media_activity', '1' );
+					if ( $sub_activity_id ) {
+						// update activity meta
+						bp_activity_update_meta( $sub_activity_id, 'bp_media_activity', '1' );
 
-							$media_args['activity_id'] = $sub_activity_id;
-						}
+						$media_args['activity_id'] = $sub_activity_id;
 					}
 				}
 
@@ -1192,10 +1187,13 @@ function bp_media_import_buddyboss_media_tables() {
 
 				if ( ! empty( $media_id ) && ! empty( $media_args['activity_id'] ) ) {
 					update_post_meta( $attachment_id, 'bp_media_activity_id', $media_args['activity_id'] );
-					update_post_meta( $attachment_id, 'bp_media_parent_activity_id', $activity_id );
 
-					if ( isset( $activity_ids[ $activity_id ] ) ) {
-						$activity_ids[ $activity_id ][] = $media_id;
+					if ( ! empty( $activity_id ) ) {
+						update_post_meta( $attachment_id, 'bp_media_parent_activity_id', $activity_id );
+
+						if ( isset( $activity_ids[ $activity_id ] ) ) {
+							$activity_ids[ $activity_id ][] = $media_id;
+						}
 					}
 
 					$imported_media_ids[] = $media_id;
@@ -1245,6 +1243,8 @@ function bp_media_import_buddyboss_forum_media() {
 
 	if ( ! empty( $forums_media_query->found_posts ) ) {
 
+		$imported_forum_media_ids = get_option( 'bp_media_import_forum_media_ids', array() );
+
 		update_option( 'bp_media_import_forums_total', $forums_media_query->found_posts );
 
 		if ( ! empty( $forums_media_query->posts ) ) {
@@ -1275,8 +1275,12 @@ function bp_media_import_buddyboss_forum_media() {
 
 							//save media is saved in attachment
 							update_post_meta( $attachment_id, 'bp_media_saved', true );
+
+							$imported_forum_media_ids[] = $media_id;
 						}
 					}
+
+					update_option( 'bp_media_import_forum_media_ids', $imported_forum_media_ids );
 
 					$media_ids = implode( ',', $media_ids );
 
@@ -1323,6 +1327,8 @@ function bp_media_import_buddyboss_topic_media() {
 
 	if ( ! empty( $topics_media_query->found_posts ) ) {
 
+		$imported_topic_media_ids = get_option( 'bp_media_import_topic_media_ids', array() );
+
 		update_option( 'bp_media_import_topics_total', $topics_media_query->found_posts );
 
 		if ( ! empty( $topics_media_query->posts ) ) {
@@ -1353,8 +1359,12 @@ function bp_media_import_buddyboss_topic_media() {
 
 							//save media is saved in attachment
 							update_post_meta( $attachment_id, 'bp_media_saved', true );
+
+							$imported_topic_media_ids[] = $media_id;
 						}
 					}
+
+					update_option( 'bp_media_import_topic_media_ids', $imported_topic_media_ids );
 
 					$media_ids = implode( ',', $media_ids );
 
@@ -1401,6 +1411,8 @@ function bp_media_import_buddyboss_reply_media() {
 
 	if ( ! empty( $replies_media_query->found_posts ) ) {
 
+		$imported_reply_media_ids = get_option( 'bp_media_import_reply_media_ids', array() );
+
 		update_option( 'bp_media_import_replies_total', $replies_media_query->found_posts );
 
 		if ( ! empty( $replies_media_query->posts ) ) {
@@ -1431,8 +1443,12 @@ function bp_media_import_buddyboss_reply_media() {
 
 							//save media is saved in attachment
 							update_post_meta( $attachment_id, 'bp_media_saved', true );
+
+							$imported_reply_media_ids[] = $media_id;
 						}
 					}
+
+					update_option( 'bp_media_import_reply_media_ids', $imported_reply_media_ids );
 
 					$media_ids = implode( ',', $media_ids );
 
@@ -1464,21 +1480,23 @@ function bp_media_import_reset_media_albums() {
 	$bp_media_table        = $wpdb->prefix . 'bp_media';
 	$bp_media_albums_table = $wpdb->prefix . 'bp_media_albums';
 
-	$albums = $wpdb->get_results( "SELECT DISTINCT id FROM {$bp_media_albums_table}" );
+	$album_ids = get_option( 'bp_media_import_albums_ids', array() );
 
-	if ( ! empty( $albums ) ) {
-		foreach ( $albums as $album ) {
+	remove_action( 'bp_activity_after_delete', 'bp_media_delete_activity_media' );
 
-			if ( empty( $album->id ) ) {
+	if ( ! empty( $album_ids ) ) {
+		foreach ( $album_ids as $new_album_id => $old_album_id ) {
+
+			if ( empty( $new_album_id ) ) {
 				continue;
 			}
 
-			$album_obj = new BP_Media_Album( $album->id );
+			$album_obj = new BP_Media_Album( $new_album_id );
 
 			if ( ! empty( $album_obj->id ) ) {
-				$media_ids = BP_Media::get_album_media_ids( $album->id );
+				$media_ids = BP_Media::get_album_media_ids( $album_obj->id );
 				if ( ! empty( $media_ids ) ) {
-					foreach( $media_ids as $media ) {
+					foreach ( $media_ids as $media ) {
 						$media_obj = new BP_Media( $media );
 
 						if ( ! empty( $media_obj->activity_id ) && bp_is_active( 'activity' ) ) {
@@ -1496,8 +1514,9 @@ function bp_media_import_reset_media_albums() {
 
 								// Deleting an activity.
 							} else {
-								if ( bp_activity_delete( array( 'id'      => $activity->id,
-								                                'user_id' => $activity->user_id
+								if ( bp_activity_delete( array(
+									'id'      => $activity->id,
+									'user_id' => $activity->user_id
 								) ) ) {
 									/** This action is documented in bp-activity/bp-activity-actions.php */
 									do_action( 'bp_activity_action_delete_activity', $activity->id, $activity->user_id );
@@ -1507,13 +1526,19 @@ function bp_media_import_reset_media_albums() {
 					}
 
 					$media_ids = implode( ',', $media_ids );
-					$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$media_ids});" );
+					if( ! empty( $media_ids ) ) {
+						$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$media_ids});" );
+					}
 				}
 			}
 
 			$wpdb->query( "DELETE FROM {$bp_media_albums_table} WHERE id = {$album_obj->id};" );
 		}
 	}
+
+	add_action( 'bp_activity_after_delete', 'bp_media_delete_activity_media' );
+
+	update_option( 'bp_media_import_status', 'reset_media' );
 }
 
 /**
@@ -1524,19 +1549,21 @@ function bp_media_import_reset_media_albums() {
 function bp_media_import_reset_media() {
 	global $wpdb;
 
-	$bp_media_table        = $wpdb->prefix . 'bp_media';
+	$bp_media_table = $wpdb->prefix . 'bp_media';
 
-	$medias = $wpdb->get_results( "SELECT DISTINCT id FROM {$bp_media_table}" );
+	$medias = get_option( 'bp_media_import_media_ids', array() );
+
+	remove_action( 'bp_activity_after_delete', 'bp_media_delete_activity_media' );
 
 	if ( ! empty( $medias ) ) {
 		$media_ids = array();
-		foreach( $medias as $media ) {
+		foreach ( $medias as $media ) {
 
-			if ( empty( $media->id ) ) {
+			if ( empty( $media ) ) {
 				continue;
 			}
 
-			$media_obj = new BP_Media( $media->id );
+			$media_obj = new BP_Media( $media );
 
 			if ( ! empty( $media_obj->activity_id ) && bp_is_active( 'activity' ) ) {
 				$activity = new BP_Activity_Activity( (int) $media_obj->activity_id );
@@ -1553,8 +1580,9 @@ function bp_media_import_reset_media() {
 
 					// Deleting an activity.
 				} else {
-					if ( bp_activity_delete( array( 'id'      => $activity->id,
-					                                'user_id' => $activity->user_id
+					if ( bp_activity_delete( array(
+						'id'      => $activity->id,
+						'user_id' => $activity->user_id
 					) ) ) {
 						/** This action is documented in bp-activity/bp-activity-actions.php */
 						do_action( 'bp_activity_action_delete_activity', $activity->id, $activity->user_id );
@@ -1568,8 +1596,77 @@ function bp_media_import_reset_media() {
 		}
 
 		$media_ids = implode( ',', $media_ids );
-		$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$media_ids})" );
+		if ( ! empty( $media_ids ) ) {
+			$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$media_ids})" );
+		}
 	}
+
+	add_action( 'bp_activity_after_delete', 'bp_media_delete_activity_media' );
+
+	update_option( 'bp_media_import_status', 'reset_forum' );
+}
+
+/**
+ * Reset all media related data in forums
+ *
+ * @since BuddyBoss 1.0.10
+ */
+function bp_media_import_reset_forum_media() {
+	global $wpdb;
+
+	$bp_media_table = $wpdb->prefix . 'bp_media';
+
+	$medias = get_option( 'bp_media_import_forum_media_ids', array() );
+
+	if ( ! empty( $medias ) ) {
+
+		$medias = implode( ',', $medias );
+		$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$medias})" );
+	}
+
+	update_option( 'bp_media_import_status', 'reset_topic' );
+}
+
+/**
+ * Reset all media related data in topics
+ *
+ * @since BuddyBoss 1.0.10
+ */
+function bp_media_import_reset_topic_media() {
+	global $wpdb;
+
+	$bp_media_table = $wpdb->prefix . 'bp_media';
+
+	$medias = get_option( 'bp_media_import_topic_media_ids', array() );
+
+	if ( ! empty( $medias ) ) {
+
+		$medias = implode( ',', $medias );
+		$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$medias})" );
+	}
+
+	update_option( 'bp_media_import_status', 'reset_reply' );
+}
+
+/**
+ * Reset all media related data in topics
+ *
+ * @since BuddyBoss 1.0.10
+ */
+function bp_media_import_reset_reply_media() {
+	global $wpdb;
+
+	$bp_media_table = $wpdb->prefix . 'bp_media';
+
+	$medias = get_option( 'bp_media_import_reply_media_ids', array() );
+
+	if ( ! empty( $medias ) ) {
+
+		$medias = implode( ',', $medias );
+		$wpdb->query( "DELETE FROM {$bp_media_table} WHERE id IN ({$medias})" );
+	}
+
+	update_option( 'bp_media_import_status', 'reset_options' );
 }
 
 /**
@@ -1578,7 +1675,6 @@ function bp_media_import_reset_media() {
  * @since BuddyBoss 1.0.5
  */
 function bp_media_import_reset_options() {
-	update_option( 'bp_media_import_status', '' );
 	update_option( 'bp_media_import_total_media', 0 );
 	update_option( 'bp_media_import_total_albums', 0 );
 	update_option( 'bp_media_import_albums_done', 0 );
@@ -1589,4 +1685,112 @@ function bp_media_import_reset_options() {
 	update_option( 'bp_media_import_forums_total', 0 );
 	update_option( 'bp_media_import_topics_total', 0 );
 	update_option( 'bp_media_import_replies_total', 0 );
+	delete_option( 'bp_media_import_reply_media_ids' );
+	delete_option( 'bp_media_import_topic_media_ids' );
+	delete_option( 'bp_media_import_forum_media_ids' );
+	delete_option( 'bp_media_import_media_ids' );
+	delete_option( 'bp_media_import_albums_ids' );
+
+	update_option( 'bp_media_import_status', 'start' );
+}
+
+/**
+ * AJAX function for media import status
+ *
+ * @since BuddyBoss 1.0.0
+ */
+function bp_media_import_status_request() {
+
+	$import_status = get_option( 'bp_media_import_status' );
+
+	if ( 'reset_albums' == $import_status ) {
+		bp_media_import_reset_media_albums();
+	} else if ( 'reset_media' == $import_status ) {
+		bp_media_import_reset_media();
+	} else if ( 'reset_forum' == $import_status ) {
+		bp_media_import_reset_forum_media();
+	} else if ( 'reset_topic' == $import_status ) {
+		bp_media_import_reset_topic_media();
+	} else if ( 'reset_reply' == $import_status ) {
+		bp_media_import_reset_reply_media();
+	} else if ( 'reset_options' == $import_status ) {
+		bp_media_import_reset_options();
+	} else if ( 'start' == $import_status ) {
+
+		update_option( 'bp_media_import_status', 'importing' );
+
+		bp_media_import_buddyboss_media_tables();
+		bp_media_import_buddyboss_forum_media();
+		bp_media_import_buddyboss_topic_media();
+		bp_media_import_buddyboss_reply_media();
+	} else {
+
+		$total_media   = get_option( 'bp_media_import_total_media', 0 );
+		$total_albums  = get_option( 'bp_media_import_total_albums', 0 );
+		$albums_done   = get_option( 'bp_media_import_albums_done', 0 );
+		$media_done    = get_option( 'bp_media_import_media_done', 0 );
+		$forums_done   = get_option( 'bp_media_import_forums_done', 0 );
+		$forums_total  = get_option( 'bp_media_import_forums_total', 0 );
+		$topics_done   = get_option( 'bp_media_import_topics_done', 0 );
+		$topics_total  = get_option( 'bp_media_import_topics_total', 0 );
+		$replies_done  = get_option( 'bp_media_import_replies_done', 0 );
+		$replies_total = get_option( 'bp_media_import_replies_total', 0 );
+
+		$importing = false;
+		if ( $albums_done != $total_albums || $media_done != $total_media ) {
+			bp_media_import_buddyboss_media_tables();
+			$importing = true;
+		}
+
+		if ( bp_is_active( 'forums' ) ) {
+			if ( $forums_done != $forums_total ) {
+				bp_media_import_buddyboss_forum_media();
+				$importing = true;
+			}
+
+			if ( $topics_done != $topics_total ) {
+				bp_media_import_buddyboss_topic_media();
+				$importing = true;
+			}
+
+			if ( $replies_done != $replies_total ) {
+				bp_media_import_buddyboss_reply_media();
+				$importing = true;
+			}
+		}
+
+		if ( ! $importing ) {
+			update_option( 'bp_media_import_status', 'done' );
+		} else {
+			update_option( 'bp_media_import_status', 'importing' );
+		}
+	}
+
+	$import_status = get_option( 'bp_media_import_status' );
+	$total_media   = get_option( 'bp_media_import_total_media', 0 );
+	$total_albums  = get_option( 'bp_media_import_total_albums', 0 );
+	$albums_done   = get_option( 'bp_media_import_albums_done', 0 );
+	$media_done    = get_option( 'bp_media_import_media_done', 0 );
+	$forums_done   = get_option( 'bp_media_import_forums_done', 0 );
+	$forums_total  = get_option( 'bp_media_import_forums_total', 0 );
+	$topics_done   = get_option( 'bp_media_import_topics_done', 0 );
+	$topics_total  = get_option( 'bp_media_import_topics_total', 0 );
+	$replies_done  = get_option( 'bp_media_import_replies_done', 0 );
+	$replies_total = get_option( 'bp_media_import_replies_total', 0 );
+
+	wp_send_json_success( array(
+		'total_media'   => $total_media,
+		'total_albums'  => $total_albums,
+		'albums_done'   => $albums_done,
+		'media_done'    => $media_done,
+		'forums_done'   => $forums_done,
+		'topics_done'   => $topics_done,
+		'replies_done'  => $replies_done,
+		'forums_total'  => $forums_total,
+		'topics_total'  => $topics_total,
+		'replies_total' => $replies_total,
+		'import_status' => $import_status,
+		'success_msg'   => __( 'BuddyBoss Media data update is complete! Any previously uploaded member photos should display in their profiles now.', 'buddyboss' ),
+		'error_msg'     => __( 'BuddyBoss Media data update is failing!', 'buddyboss' ),
+	) );
 }
